Refactor: Restructure project by moving scripts to a dedicated folder and removing unused files
- Deleted `src/types.ts` and `src/utils.ts` files.
- Created `src/scripts/types.ts` and `src/scripts/utils.ts` for type definitions and utility functions.
- Moved core functionality from `newtab.ts` to `newtab-core.ts` for better testability.
- Updated import paths in test files to reflect new structure.
- Enhanced favicon management and bookmark processing logic.
- Improved event handling for folder clicks and bookmark rendering.
