+++ src/newtab.html
@@ -22,6 +22,6 @@
         </main>
     </div>
     
-    <script type="module" src="newtab.js"></script>
+    <script type="module" src="scripts/newtab.js"></script>
 </body>
 </html>
