+++ src/scripts/newtab-core.ts
