+++ src/scripts/newtab.ts
